+++ src/main/resources/application-dev.yml
@@ -1,10 +1,12 @@
 spring:
+  # mysql 설정
   datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
     url: jdbc:mysql://172.21.138.60:3306/idealstudy
-#    url: jdbc:mysql://172.17.129.27:3306/idealstudy (성진 wsl 주소)
-    username: manager
-    password: manager
+    #    url: jdbc:mysql://172.17.129.27:3306/idealstudy (성진 wsl 주소)
+  # redis 설정
+  data:
+    redis:
+      host: 172.21.138.60
 
 server:
   backend-domain-url: http://localhost:8080
